+++ db-viewer-ui/src/Components/ObjectFilter/ObjectFilter.tsx
@@ -2,7 +2,7 @@ import { TimeUtils, StringUtils } from "@skbkontur/edi-ui";
 import { ColumnStack, Fit, RowStack } from "@skbkontur/react-stack-layout";
 import { Input } from "@skbkontur/react-ui";
 import { tooltip, ValidationInfo, ValidationWrapper } from "@skbkontur/react-ui-validations";
-import React from "react";
+import type { ReactElement } from "react";
 
 import { Condition } from "../../Domain/Api/DataTypes/Condition";
 import { ObjectFieldFilterOperator } from "../../Domain/Api/DataTypes/ObjectFieldFilterOperator";
@@ -19,9 +19,8 @@ interface ObjectFilterProps {
     tableColumns: PropertyMetaInformation[];
 }
 
-export class ObjectFilter extends React.Component<ObjectFilterProps> {
-    public getCondition(property: PropertyMetaInformation): Condition {
-        const { conditions } = this.props;
+export const ObjectFilter = ({ conditions, onChange, tableColumns }: ObjectFilterProps): ReactElement => {
+    const getCondition = (property: PropertyMetaInformation): Condition => {
         const result = conditions.find(x => x.path === property.name);
         if (result == null) {
             return {
@@ -31,10 +30,9 @@ export class ObjectFilter extends React.Component<ObjectFilterProps> {
             };
         }
         return result;
-    }
+    };
 
-    public updateItem(property: PropertyMetaInformation, conditionUpdate: Partial<Condition>) {
-        const { conditions, onChange } = this.props;
+    const updateItem = (property: PropertyMetaInformation, conditionUpdate: Partial<Condition>) => {
         const conditionIndex = conditions.findIndex(x => x.path === property.name);
         if (conditionIndex >= 0) {
             onChange([
@@ -43,11 +41,18 @@ export class ObjectFilter extends React.Component<ObjectFilterProps> {
                 ...conditions.slice(conditionIndex + 1),
             ]);
         } else {
-            onChange([...conditions, { ...this.getCondition(property), ...conditionUpdate }]);
+            onChange([...conditions, { ...getCondition(property), ...conditionUpdate }]);
         }
-    }
+    };
 
-    public renderProperty(property: PropertyMetaInformation, value: Nullable<string>): React.ReactElement {
+    const getValidation = (property: PropertyMetaInformation, value: Nullable<string>): ValidationInfo | null => {
+        if (property.isRequired && StringUtils.isNullOrWhitespace(value)) {
+            return { message: "Поле должно быть заполнено", type: "submit" };
+        }
+        return validateObjectField(value);
+    };
+
+    const renderProperty = (property: PropertyMetaInformation, value: Nullable<string>): ReactElement => {
         const type = property.type.typeName;
         if (type === "DateTime" || type === "DateTimeOffset") {
             return (
@@ -58,7 +63,7 @@ export class ObjectFilter extends React.Component<ObjectFilterProps> {
                             defaultTime={""}
                             error={false}
                             timeZone={TimeUtils.TimeZones.UTC}
-                            onChange={date => this.updateItem(property, { value: TimeUtils.timestampToTicks(date) })}
+                            onChange={date => updateItem(property, { value: TimeUtils.timestampToTicks(date) })}
                             value={value ? TimeUtils.ticksToTimestamp(value) : null}
                         />
                     </Fit>
@@ -66,11 +71,11 @@ export class ObjectFilter extends React.Component<ObjectFilterProps> {
                         <ValidationWrapper
                             data-tid="DateTimeValidation"
                             renderMessage={tooltip("right middle")}
-                            validationInfo={this.getValidation(property, value)}>
+                            validationInfo={getValidation(property, value)}>
                             <Input
                                 mask="999999999999999999"
                                 data-tid={"DateTimeInTicks"}
-                                onValueChange={nextValue => this.updateItem(property, { value: nextValue })}
+                                onValueChange={nextValue => updateItem(property, { value: nextValue })}
                                 value={value ? value : ""}
                             />
                         </ValidationWrapper>
@@ -83,12 +88,12 @@ export class ObjectFilter extends React.Component<ObjectFilterProps> {
                 <ValidationWrapper
                     data-tid="BooleanValidation"
                     renderMessage={tooltip("right middle")}
-                    validationInfo={this.getValidation(property, value)}>
+                    validationInfo={getValidation(property, value)}>
                     <StyledSelect
                         data-tid="BooleanSelect"
                         items={[null, "true", "false"].map(x => [x, String(x)])}
                         onValueChange={(nextValue: any) => {
-                            this.updateItem(property, { value: nextValue });
+                            updateItem(property, { value: nextValue });
                         }}
                         value={value || undefined}
                     />
@@ -100,12 +105,12 @@ export class ObjectFilter extends React.Component<ObjectFilterProps> {
                 <ValidationWrapper
                     data-tid="EnumValidation"
                     renderMessage={tooltip("right middle")}
-                    validationInfo={this.getValidation(property, value)}>
+                    validationInfo={getValidation(property, value)}>
                     <StyledSelect
                         data-tid="EnumSelect"
                         items={[null, ...property.availableValues].map(x => [x, String(x)])}
                         onValueChange={(nextValue: any) => {
-                            this.updateItem(property, { value: nextValue });
+                            updateItem(property, { value: nextValue });
                         }}
                         value={value || undefined}
                     />
@@ -116,47 +121,34 @@ export class ObjectFilter extends React.Component<ObjectFilterProps> {
             <ValidationWrapper
                 data-tid="InputValidation"
                 renderMessage={tooltip("right middle")}
-                validationInfo={this.getValidation(property, value)}>
+                validationInfo={getValidation(property, value)}>
                 <Input
                     data-tid="Input"
-                    onValueChange={nextValue => this.updateItem(property, { value: nextValue })}
+                    onValueChange={nextValue => updateItem(property, { value: nextValue })}
                     value={value || ""}
                 />
             </ValidationWrapper>
         );
-    }
+    };
 
-    public getValidation(property: PropertyMetaInformation, value: string | null | undefined): ValidationInfo | null {
-        if (property.isRequired && StringUtils.isNullOrWhitespace(value)) {
-            return { message: "Поле должно быть заполнено", type: "submit" };
-        }
-        return validateObjectField(value);
-    }
-
-    public render(): React.ReactElement {
-        const { tableColumns } = this.props;
-
-        return (
-            <ColumnStack gap={2} data-tid="ObjectFilters">
-                {tableColumns
-                    .map(x => [x, this.getCondition(x)] as [PropertyMetaInformation, Condition])
-                    .map(([property, condition]) => (
-                        <FormRow key={property.name} captionWidth={200} caption={property.name} data-tid="Filter">
-                            <RowStack baseline block gap={5} data-tid={property.name}>
-                                <Fit>
-                                    <OperatorSelect
-                                        value={condition.operator}
-                                        onChange={value => this.updateItem(property, { operator: value })}
-                                        availableValues={
-                                            property.availableFilters || [ObjectFieldFilterOperator.Equals]
-                                        }
-                                    />
-                                </Fit>
-                                <Fit>{this.renderProperty(property, condition.value)}</Fit>
-                            </RowStack>
-                        </FormRow>
-                    ))}
-            </ColumnStack>
-        );
-    }
-}
+    return (
+        <ColumnStack gap={2} data-tid="ObjectFilters">
+            {tableColumns
+                .map(x => [x, getCondition(x)] as [PropertyMetaInformation, Condition])
+                .map(([property, condition]) => (
+                    <FormRow key={property.name} captionWidth={200} caption={property.name} data-tid="Filter">
+                        <RowStack baseline block gap={5} data-tid={property.name}>
+                            <Fit>
+                                <OperatorSelect
+                                    value={condition.operator}
+                                    onChange={value => updateItem(property, { operator: value })}
+                                    availableValues={property.availableFilters || [ObjectFieldFilterOperator.Equals]}
+                                />
+                            </Fit>
+                            <Fit>{renderProperty(property, condition.value)}</Fit>
+                        </RowStack>
+                    </FormRow>
+                ))}
+        </ColumnStack>
+    );
+};